added acceptance test for user using blackbox random port.

+++ episodic-events/src/main/java/com/example/episodicevents/DataDefault.java
@@ -4,5 +4,5 @@ import lombok.Data;
 
 @Data
 public class DataDefault {
-  private Long offset;
+  private int offset;
 }
